ai-service/requirements.txt
# ── Curezy AURANET — Production Requirements ────────────────────────────────
# Core API
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
python-multipart>=0.0.9
slowapi>=0.1.9

# AI / LLM
ollama>=0.2.0
runpod>=1.6.0
groq>=0.5.0

# Data & utilities
python-dotenv>=1.0.0
httpx>=0.27.0
aiohttp>=3.9.0

# Database
supabase>=2.3.0

# Medical / NLP (Optional on Render, Mandatory on RunPod)
# spacy>=3.7.0  <-- removed to fix Render build
langchain-core>=0.1.0

# Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
cryptography>=41.0.0
cffi>=1.15.0

# Document parsing
PyPDF2>=3.0.0
python-docx>=1.1.0

# Image analysis (X-ray)
Pillow>=10.0.0
numpy>=1.26.0,<2.0.0

# Excel export (benchmarks)
openpyxl>=3.1.0
pytesseract>=0.3.10

# Missing dependencies for RunPod deployment
pydicom>=2.4.4
opencv-python-headless>=4.9.0.80
# scispacy>=0.5.4 <-- removed to fix Render build
resend>=0.8.0

# Semantic Caching
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
